fix: memory leak on simulate request, graceful shutdown
Previously, there was no timeout for RPC simulation calls, which caused a memory leak due to requests that were left hanging indefinitely. This fix introduces a timer to ensure requests complete in a timely manner.
As for graceful shutdown, we had complex shutdown logic that did not work anyway, I replaced it with context conselation propagation through the pipeline.
In addition, added timeout for aux commands.